[#218] First cut of JSON plugin driven updates on /seeithappen
Add json_data_render_update() to functions.php. This function renders
one instance of a project update box.

Add Akvo-responsive/json_data_templates. This folder is used to store
the code snippets uesed in the JSON plugin template form fields.
A snippet should include all information needed to instantiate it using
the JSON plugin.

Add Akvo-responsive/json_data_templates/rsr-updates.php. This is the
code snippet that is to be used in the JSON plugin. The API resource is
still missing a couple of features, so there's some bogus data in the
call to json_data_render_update()