[#218] Use JSON feed on /seeithappen/all-rsr-project-updates page
Add json_data_templates/rsr-many-updates.template.php to be used as the
template for the JSON feed.

Replace the javascript implementation of the updates rendering with the
feed template.

Remove the unused javascript.

The page now displays up to 60 updates in one go. There is no point in
fetching them incrementally. Since we currently filter out updates
without an image not all of the 60 fetched may be displayed.
When the RSR API resource is fixed it will be possible to request only
updates with images, guaranteeing a cerain number of updates.