[#223] Render all RSR projects on a google map
Use the JSON plugin to fetch the primary location data for all projects
in RSR. Use this to render a global map similar to the one in RSR:
http://rsr.akvo.org/maps/projects/all/.

The changes in networkPage.php is a modified version of the code used in
RSR to render the map. Note that there are two calls to do_shorcode().
This is because the RSR API will not hand over more than 1000 objects
in one go and there are currently over 1700 project in RSR.

Add Akvo-responsive/json_data_templates/rsr-project-global-map.template.php
that massages the JSON into a format easily used in the networkPage.php.
The main reason for doing this in PHP rather than in the client
javascript is that I found it really hard to handle the case of missing
objects in javascript.