[#587] Remove the RSR projects map from /seeithappen
This is because of the performance problems in RSR leading to a rate
limit of 10 objects.

The current JSON plugin only caches "all" data it pulls, it has no
concept of a permanent cache that updates with new data.

So we have to remove the map until we can spec the new JSON plugin with
long term cache management.